Announcement Details fix
- Poprawka kolorów,
- Poprawka wyglądu tile'ów aplikacji
- Margin na obrazek
- Poprawka szerokości kolumn grida
- Poprawka wyglądu Pet Details i Shelter Details, dodane row-gap
- Dodane ikonki oznaczające numer telefonu i maila (react-icons)
- Announcement Details podzielony na pliki
