+++ src/styles/global.ts
@@ -1,4 +1,4 @@
-import { createGlobalStyle } from 'styled-components';
+import styled, { createGlobalStyle } from 'styled-components';
 
 const GlobalStyle = createGlobalStyle`
   body {
@@ -12,5 +12,23 @@ const GlobalStyle = createGlobalStyle`
     box-sizing: border-box;
   }
 `;
+
+export const Title = styled.h1`
+margin: 0;
+padding: 5px;
+font-size: 25px;
+`
+
+export const TextDetails = styled.p`
+  margin: 0;
+  padding: 2px;
+  font-size: 15px;
+`
+
+export const DescriptionText = styled.p`
+  margin: 0;
+  padding: 5px;
+  font-size: 14px;
+`
 
 export default GlobalStyle;
