+++ src/styles/theme.ts
@@ -2,6 +2,8 @@ export const defaultTheme = {
   colors: {
     powderWhite: "#FFFDF9",
     background: "#FF00FF",
+    tomato: "#ff6347",
+    darkLemonLime: "#79D021",
     main: "rgba(62,170,111, 1)"
   },
   fonts: ["sans-serif", "Roboto"],
